Split amp-selector component files
