+++ extensions/amp-selector/1.0/component.css.js
