+++ extensions/amp-selector/1.0/component.type.js
