+++ extensions/amp-selector/1.0/storybook/Basic.js
@@ -15,7 +15,7 @@
  */
 
 import * as Preact from '../../../../src/preact';
-import {Option, Selector} from '../selector';
+import {Option, Selector} from '../component';
 import {select, withKnobs} from '@storybook/addon-knobs';
 import {useState} from '../../../../src/preact';
 import {withA11y} from '@storybook/addon-a11y';
